7 - Export modules at entry file index and reorganize file dir paths
