+++ example/chapters/intro.js
@@ -1,5 +1,5 @@
-import { Chapter } from '../../src/chapter.js';
-import { Scene } from '../../src/scene.js';
+import { Chapter } from '../../index.js';
+import { Scene } from '../../index.js';
 
 const enterMainCharacter = new Scene('input', 'What is the name of your character?');
 const chooseDifficultyLevel = new Scene('choice', 'What level of difficulty you want?', ['Easy', 'Medium', 'Hard']);
