+++ index.js
@@ -1,0 +1,3 @@
+export { Chapter } from './src/chapter.js';
+export { Scene } from './src/scene.js';
+export { GameEngine } from './src/game-engine.js';
