+++ package.json
@@ -17,7 +17,7 @@
   "license": "ISC",
   "author": "Joshua Cadavez",
   "type": "module",
-  "main": "./example/index.js",
+  "main": "./index.js",
   "scripts": {
     "lint": "eslint .",
     "start:example": "node ./example/index.js",
